Functional PersistentList, interface troubles.
This commit adds the majority of the remaining PersistentList code. A
few unpleasant lessons about the different inheritance models for both
languages along the way, but at least everything compiles.

Some notes about this can be found in the developer notes. Copied here
for posterity:

The lack of union types in Go presents a significant problem when it
comes to interface management. Consider the case of ISeq and
IPersistentCollection - two interfaces that specify a different
function signature for the same function, `cons`. By itself this isn't
necessarily a problem, but it is when we come to the case of the
EmptyList class/struct, which must implement both (it inherits the
IPersisentCollection implementation from PersistentList). I
believe the best choice for the moment is to refactor both interfaces
to have independent functions (e.g. ConsISeq and
ConsIPersistentCollection) and target those directly. In the
future I expect a general `Cons` will be required with type switching
at runtime but this will have to do for now.