Initial work on Murmur3
This commit adds a Murmur3.go file for working with murmurhash3. It took
a few hours to figure out since the default interface for hashing in Go
is to return a byte array, not an integer, and I wanted to make sure the
damn thing was actually working. It seems to be, though, so I can now
take my reference implementation for integers and do whatever the hell I
feel like later.