Various additions, fixed two bugs
Bugs:
 - The Symbol String() function was attempting to validate the memory address of
 a value, which had some unexpected behaviors that caused all string
 representations to be the empty string. In the future, it might make sense to
 have the Symbol `ns` and `name` attributes be pointers to strings rather than
 strings themselves, but I don't have a firm decision on this either way.
 - The PersistentArrayMap constructor had a bug that initialized an inner and
 outer loop to the same value, which were then compared against each other and
 immediately threw an exception.
 - There is currently some sort of bug with the locking mechanism for
 Namespaces.

As a general note, running `go build` is insufficient to find bugs. Because
there are many data structures that are actually initialized in order for the
compiler itself to work, running `go test` is an effective strategy here whether
or not the tests themselves actually cover the relevant behavior.