Validated persistenthashmap functionality
This commit includes a ton of functionality for persistenthashmaps. Up until now
none of the persistenthashmap code was tested in any way; this commit adds a ton
of function implementations, particularly around the various node types, and has
a basic test for reading larger maps that doesn't cause the whole program to
instantly die.